Get authStrategy working
