+++ middleware/auth.ts
@@ -1,8 +1,13 @@
 // Following example: https://github.com/atinux/atidone/blob/main/app/middleware/auth.ts
 export default defineNuxtRouteMiddleware(async (to) => {
   const { loggedIn } = useUserSession();
+  const config = useRuntimeConfig();
 
-  if (!loggedIn.value && to.path !== "/login") {
+  if (
+    config.public.authStrategy === "auth0" &&
+    !loggedIn.value &&
+    to.path !== "/login"
+  ) {
     return navigateTo("/login");
   }
 });
